Update: include redirection url passed as a prop
In the previous version the redirection was hardcoded
to `settings/general`.

For reusability, here we allow a redirect link to be
passed as a prop. The default fallback url is set to
`/settings/general`.

Relevant names are changed to remove references
to the `general` section.

README of the `redirectNonJetpack` component is updated
to include the relevant changes.